+++ Oxide.Ext.RustApi/Interfaces/IAuthenticationService.cs
@@ -1,4 +1,5 @@
-﻿using Oxide.Ext.RustApi.Models;
+﻿using System.Net;
+using Oxide.Ext.RustApi.Models;
 
 namespace Oxide.Ext.RustApi.Interfaces
 {
@@ -10,11 +11,9 @@ namespace Oxide.Ext.RustApi.Interfaces
         /// <summary>
         /// Test if sign is valid.
         /// </summary>
-        /// <param name="sign">Current sign in request.</param>
-        /// <param name="route">Route url value.</param>
-        /// <param name="requestContent">Content data.</param>
+        /// <param name="context">Http request context.</param>
         /// <param name="userInfo">Found user info.</param>
         /// <returns></returns>
-        bool TryToGetUser(string user, string sign, string route, string requestContent, out ApiUserInfo userInfo);
+        bool TryToGetUser(HttpListenerContext context, out ApiUserInfo userInfo);
     }
 }
